add error handler, notifications, and alert

+++ script.js
@@ -11,12 +11,14 @@ $(document).ready(function () {
   $('#reg-btn').click(function (e) {
     e.preventDefault();
     $('#login-container').hide();
+    $('#reg-alert').hide();
     $('#reg-container').show();
   });
 
   $('#cancel-btn').click(function (e) {
     e.preventDefault();
     auth();
+    $('#login-form')[0].reset();
   });
 
   $('#reg-form').submit(function (e) {
@@ -28,11 +30,15 @@ $(document).ready(function () {
     e.preventDefault();
     localStorage.removeItem('access_token');
     auth();
+    $('#login-form')[0].reset();
   });
 
   $('#add-btn').click(function (e) {
     e.preventDefault();
     $('#add-container').show();
+    $('#due_date').attr('min', new Date().toISOString().split('T')[0]);
+    $('#todo-alert').hide();
+    $('#add-alert').hide();
   });
 
   $('#cancel-add-btn').click(function (e) {
@@ -43,7 +49,6 @@ $(document).ready(function () {
   $('#add-form').submit(function (e) {
     e.preventDefault();
     addTodo();
-    $('#add-container').hide();
   });
 
   $('#todos').on('click', '.edit-btn', function (e) {
@@ -51,6 +56,8 @@ $(document).ready(function () {
     const id = $(this).data().id;
     showEditTodo(id);
     $('#edit-container').show();
+    $('#edit-alert').hide();
+    $('#due_date-edit').attr('min', new Date().toISOString().split('T')[0]);
     $(window).scrollTop(0);
     $('#submit-edit-btn').attr('data-id', id);
   });
@@ -59,7 +66,6 @@ $(document).ready(function () {
     e.preventDefault();
     const id = $('#submit-edit-btn').data().id;
     editTodo(id);
-    $('#edit-container').hide();
   });
 
   $('#cancel-edit-btn').click(function (e) {
@@ -70,42 +76,62 @@ $(document).ready(function () {
   $('#todos').on('click', '.delete-btn', function (e) {
     e.preventDefault();
     const id = $(this).data().id;
-    deleteTodo(id);
+    const title = $(this).data().title;
+    deleteTodo(id, title);
   });
 
   $('#todos').on('change', '.checkbox :checkbox', function () {
     const id = $(this).data().id;
+    const title = $(this).data().title;
     if (this.checked) {
       statusChange(id, 'done');
+      // $('#todo-alert')
+      //   .show()
+      //   .text(title + ' is done');
     } else {
       statusChange(id, 'undone');
+      // $('#todo-alert').hide();
     }
+  });
+
+  $('#download-btn').click(function (e) {
+    e.preventDefault();
+    download();
   });
 });
 
-function statusChange(id, status) {
+function download() {
   $.ajax({
-    type: 'PATCH',
-    url: baseURL + '/todos/' + id,
+    type: 'GET',
+    url: baseURL + '/todos/export',
     headers: {
       access_token: localStorage.getItem('access_token'),
     },
-    data: {
-      status,
+    xhrFields: {
+      responseType: 'blob',
     },
-  })
-    .done((res) => {
-      getTodos();
-    })
-    .fail((err) => {
-      console.log(err);
-    });
+    success: function (response, status, xhr) {
+      var a = document.createElement('a');
+      var url = window.URL.createObjectURL(response);
+      a.href = url;
+      a.download = 'todo.xlsx';
+      a.click();
+      window.URL.revokeObjectURL(url);
+    },
+  });
+  // .done((res) => {
+  //   auth();
+  //   console.log(res, 'done');
+  // })
+  // .fail((err) => console.log(err, 'err'));
 }
 
 function auth() {
   if (localStorage.getItem('access_token')) {
     $('#login-container').hide();
     $('#reg-container').hide();
+    $('#no-todo').hide();
+    $('#todo-alert').hide();
     $('#todos-container').show();
     $('#logout').show();
     $('#add-container').hide();
@@ -114,6 +140,8 @@ function auth() {
     getTodos();
   } else {
     $('#login-container').show();
+    $('#success-alert').hide();
+    $('#error-alert').hide();
     $('#reg-container').hide();
     $('#todos-container').hide();
     $('#add-container').hide();
@@ -131,25 +159,33 @@ function getTodos() {
     },
   })
     .done(({ data }) => {
+      if (data.length == 0) {
+        $('#todos-container').hide();
+        $('#add-container').show();
+        $('#due_date').attr('min', new Date().toISOString().split('T')[0]);
+        $('#add-alert').hide();
+        $('#no-todo').show();
+        $('#cancel-add-btn').hide();
+      }
       $('#todos').empty();
       data.forEach((el) => {
         const status = el.status == 'done' ? 'checked' : '';
-        const muted = el.status == 'done' ? 'text-muted' : '';
         const grey = el.status == 'done' ? 'bg-secondary' : '';
+        const white = el.status == 'done' ? 'text-white' : '';
         $('#todos').append(
           `<div class="list-group-item ${grey}">
             <form class="checkbox">
-              <input data-id="${el.id}" class="form-check-input me-1" type="checkbox" ${status} value="" />
-              <label>${el.status}</label>
+              <input data-id="${el.id}" data-title="${el.title}" class="form-check-input me-1" type="checkbox" ${status} value="" />
+              <label class="${white}">${el.status}</label>
             </form>
               <div class="d-flex w-100 justify-content-between">
-                <h5 class="mb-1">${el.title}</h5>
-                <small class="text-muted">3 days ago</small>
+                <h5 class="mb-1 ${white}">${el.title}</h5>
+                <small class="text-muted">${dayCount(new Date(el.due_date), new Date())} day(s) left</small>
               </div>
-              <p class="mb-1">${el.description}</p>
-              <small>Due date: ${new Date(el.due_date).toISOString().split`T`[0]}</small><br />
+              <p class="mb-1 ${white}">${el.description}</p>
+              <small class="${white}">Due date: ${new Date(el.due_date).toISOString().split`T`[0]}</small><br />
               <button type="button" class="btn btn-sm btn-primary edit-btn" data-id="${el.id}">Edit</button>
-              <button type="button" class="btn btn-sm btn-danger delete-btn" data-id="${el.id}">Delete</button>
+              <button type="button" class="btn btn-sm btn-danger delete-btn" data-id="${el.id}" data-title="${el.title}">Delete</button>
           </div>`
         );
       });
@@ -157,6 +193,11 @@ function getTodos() {
     .fail((err) => {
       console.log(err);
     });
+}
+
+function dayCount(due_date, today) {
+  const oneDay = 24 * 60 * 60 * 1000;
+  return Math.round(Math.abs((due_date - today) / oneDay));
 }
 
 function login() {
@@ -173,7 +214,7 @@ function login() {
       auth();
     })
     .fail((err) => {
-      console.log(err);
+      $('#error-alert').show().text(err.responseJSON.message);
     });
 }
 
@@ -188,9 +229,10 @@ function register() {
   })
     .done((res) => {
       auth();
+      $('#success-alert').show().text('Register success\nYou can login now!');
     })
     .fail((err) => {
-      console.log(err);
+      $('#reg-alert').show().text(err.responseJSON.message);
     });
 }
 
@@ -209,9 +251,14 @@ function addTodo() {
   })
     .done((res) => {
       getTodos();
+      $('#add-container').hide();
+      $('#todos-container').show();
+      $('#todo-alert')
+        .show()
+        .text('New todo: "' + res.data.title + '" added successfully');
     })
     .fail((err) => {
-      console.log(err);
+      $('#add-alert').show().text(err.responseJSON.message);
     });
 }
 
@@ -248,18 +295,56 @@ function editTodo(id) {
   })
     .done((res) => {
       getTodos();
+      $('#edit-container').hide();
+      $('#todo-alert')
+        .show()
+        .text('Todo: "' + res.data.title + '" updated successfully');
     })
     .fail((err) => {
-      console.log(err);
+      $('#edit-alert').show().text(err.responseJSON.message);
     });
 }
 
-function deleteTodo(id) {
+function deleteTodo(id, title) {
+  swal({
+    title: `Are you sure you want to delete "${title}" from todo?`,
+    text: 'Once deleted, you will not be able to recover this todo!',
+    icon: 'warning',
+    buttons: true,
+    dangerMode: true,
+  }).then((willDelete) => {
+    if (willDelete) {
+      $.ajax({
+        type: 'DELETE',
+        url: baseURL + '/todos/' + id,
+        headers: {
+          access_token: localStorage.getItem('access_token'),
+        },
+      })
+        .done((res) => {
+          swal(title + ' has been deleted!', {
+            icon: 'success',
+          });
+          getTodos();
+        })
+        .fail((err) => {
+          console.log(err);
+        });
+    } else {
+      // swal('Your imaginary file is safe!');
+    }
+  });
+}
+
+function statusChange(id, status) {
   $.ajax({
-    type: 'DELETE',
+    type: 'PATCH',
     url: baseURL + '/todos/' + id,
     headers: {
       access_token: localStorage.getItem('access_token'),
+    },
+    data: {
+      status,
     },
   })
     .done((res) => {
